refactor: test cases
